I am now working on fixing the data display issues on the Villains, Shorts, and Adapted Works pages.
